requirements.txt
requests>=2.28.0
typer>=0.9.0
rich>=13.0.0

# FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Database dependencies
PyYAML>=6.0



# Additional
python-dotenv>=1.0.0
